parseDouble() for increased precision

+++ src/main/java/axoloti/realunits/DecayTime.java
@@ -53,7 +53,7 @@ public class DecayTime implements NativeToReal {
             double num, mul = 1.0;
 
             try {
-                num = Float.parseFloat(matcher.group("num"));
+                num = Double.parseDouble(matcher.group("num"));
             } catch (java.lang.NumberFormatException ex) {
                 throw new ParseException("Not DecayTime", 0);
             }
