+++ src/main/java/axoloti/realunits/LinDB.java
@@ -53,7 +53,7 @@ public class LinDB implements NativeToReal {
             double num;
 
             try {
-                num = Float.parseFloat(matcher.group("num"));
+                num = Double.parseDouble(matcher.group("num"));
             } catch (java.lang.NumberFormatException ex) {
                 throw new ParseException("Not LinDB", 0);
             }
