+++ src/main/java/axoloti/realunits/SquareDB.java
@@ -53,7 +53,7 @@ public class SquareDB implements NativeToReal {
             double num;
 
             try {
-                num = Float.parseFloat(matcher.group("num"));
+                num = Double.parseDouble(matcher.group("num"));
             } catch (java.lang.NumberFormatException ex) {
                 throw new ParseException("Not SquareDB", 0);
             }
